requirements.txt
# Core Scrapy framework
scrapy==2.11.0
Twisted==23.10.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.12.1

# Data processing
pandas==2.1.3
pydantic==2.5.0
unidecode==1.3.7

# AWS SDK
boto3==1.33.5
botocore==1.33.5

# HTTP & Networking
requests==2.31.0
urllib3==2.1.0
pyOpenSSL==23.3.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3.post1

# Geocoding
geopy==2.4.1

# Monitoring
scrapeops-scrapy==0.5.4

# Proxy support (optional)
scrapy-rotating-proxies==0.6.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
Faker==20.1.0